+++ addons/gut/gut_cmdln.gd
@@ -124,14 +124,14 @@ class CmdLineParser:
 	func get_unused_options():
 		var to_return = []
 		for i in range(_opts.size()):
-			to_return.append(_opts[i])
+			to_return.append(_opts[i].split('=')[0])
 
 		var script_option = to_return.find('-s')
 		to_return.remove(script_option + 1)
 		to_return.remove(script_option)
 
 		while(_used_options.size() > 0):
-			var index = to_return.find(_used_options[0])
+			var index = to_return.find(_used_options[0].split("=")[0])
 			if(index != -1):
 				to_return.remove(index)
 			_used_options.remove(0)